Add support for calculating difference density using atom list
The code has been modified to support calculating difference density using a list of atoms instead of a structure. A new function `get_atom_list` has been added to generate this list. In the difference-density calculations, the `calculate_difference_density` function now takes an atom list as an argument instead of a model. The related adjustments to the calculation process have been applied in the C++ difference-density code.